wip check game creation
